refactor: split write record to db into files
