+++ product-service/lambda-functions/createProduct.ts
@@ -1,14 +1,7 @@
 import { APIGatewayProxyEvent } from "aws-lambda";
-import { DynamoDBDocument } from "@aws-sdk/lib-dynamodb";
-import { DynamoDB } from "@aws-sdk/client-dynamodb";
-import { randomUUID } from "crypto";
 import { sendResponse } from "./helpers/sendResponse";
-import { TProductDto, validateProductDto } from "./helpers/validateProductDto";
-
-const PRODUCTS_TABLE_NAME = process.env.PRODUCTS_TABLE_NAME || "";
-const STOCK_TABLE_NAME = process.env.STOCK_TABLE_NAME || "";
-
-const db = DynamoDBDocument.from(new DynamoDB());
+import { validateProductDto } from "./helpers/validateProductDto";
+import { writeRecordToDB } from "./helpers/writeRecordToDB";
 
 exports.handler = async (event: APIGatewayProxyEvent) => {
   console.log("ENVIRONMENT VARIABLES\n" + JSON.stringify(process.env, null, 2));
@@ -19,47 +12,19 @@ exports.handler = async (event: APIGatewayProxyEvent) => {
       message: "Invalid request, you are missing the parameter body",
     });
 
-  const productDto =
-    typeof event.body == "object" ? event.body : JSON.parse(event.body);
+  const productDto = JSON.parse(event.body);
 
-  const isValidDto = validateProductDto(productDto);
+  const isValidDto = validateProductDto(event.body);
+
   if (!isValidDto)
     return sendResponse(400, {
       message: "Invalid request body",
     });
 
-  const id = randomUUID();
-
-  const { title, description, price, count } = productDto as TProductDto;
-
-  const productsParams = {
-    TableName: PRODUCTS_TABLE_NAME,
-    Item: {
-      id,
-      title,
-      description,
-      price,
-    },
-  };
-
-  const stockParams = {
-    TableName: STOCK_TABLE_NAME,
-    Item: {
-      productId: id,
-      count,
-    },
-  };
-
-  const transactParams = {
-    TransactItems: [{ Put: productsParams }, { Put: stockParams }],
-  };
-
   try {
-    await db.transactWrite(transactParams);
-    return sendResponse(201, {
-      ...productDto,
-      id,
-    });
+    const product = await writeRecordToDB(productDto);
+
+    return sendResponse(201, product);
   } catch (dbError) {
     return sendResponse(500, { message: `DynamoDB Error: ${dbError}` });
   }
